Fix the destroy action in the comments controller

+++ app/controllers/comments_controller.rb
@@ -7,9 +7,7 @@ class CommentsController < ApplicationController
 
   def destroy
     comment = Comment.find params[:id]
-    if comment.user.id == current_user.id
-      comment.destroy
-    end
+    comment.destroy if comment.user.id == current_user.id
   end
 
   private
